feat: Add axios and jsonwebtoken dependencies
This commit adds the axios and jsonwebtoken dependencies to the project. These dependencies are required for making HTTP requests and handling JSON Web Tokens (JWT) respectively. The axios library will be used to replace the existing fetch API calls in the AuthProvider component, while the jsonwebtoken library will be used for generating and validating JWTs. This change will improve the codebase by leveraging well-established libraries for these functionalities.